v1

+++ wwwroot/Posts/js/siteUI.js
@@ -17,7 +17,7 @@ let itemLayout;
 let waiting = null;
 let showKeywords = false;
 let keywordsOnchangeTimger = null;
-
+sessionStorage.setItem("messageCode", " ");
 Init_UI();
 async function Init_UI() {
     postsPanel = new PageManager('postsScrollPanel', 'postsPanel', 'postSample', renderPosts);
@@ -162,6 +162,11 @@ function showConfirmLogin() {
     showForm();
     renderLoginConfirm();
 }
+function showEditUser() {
+    showForm();
+    renderEditAccountForm();
+    
+}
 function showUserManager() {
     showForm();
     renderUserManager();
@@ -297,6 +302,23 @@ function updateDropDownMenu() {
     let DDMenu = $("#DDMenu");
     let selectClass = selectedCategory === "" ? "fa-check" : "fa-fw";
     DDMenu.empty();
+
+    let user = JSON.parse(sessionStorage.getItem("User"));
+    
+    if (user != null && user != undefined) {
+        // Ajouter l'image et le nom de l'utilisateur connecté
+        DDMenu.append($(`
+            <div class="dropdown-item menuItemLayout" id="userProfileCmd">
+                <img src="${user.Avatar}" alt="${user.Name}" class="userProfileImage" />
+                <span class="userName">${user.Name}</span>
+            </div>
+        `));
+        DDMenu.append($(`
+            <div class="dropdown-item menuItemLayout" id="userEditCmd">
+                <i class="menuIcon fa-solid fa-user-pen"></i> Modifier votre profil
+            </div>
+        `));
+    }
     DDMenu.append($(`
         <div class="dropdown-item menuItemLayout" id="userManagerCmd">
             <i class="menuIcon fa fa-user-gear mx-2"></i> Gestion d'usagers
@@ -305,11 +327,24 @@ function updateDropDownMenu() {
     DDMenu.append($(`
         <div class="dropdown-divider"></div>
     `));
-    DDMenu.append($(`
-        <div class="dropdown-item menuItemLayout" id="loginCmd">
-            <i class="menuIcon fa fa-right-to-bracket mx-2"></i> Se Connecter
-        </div>
-    `));
+    user = JSON.parse(sessionStorage.getItem("User"));
+    if(user == null || user == undefined)
+    {
+        DDMenu.append($(`
+            <div class="dropdown-item menuItemLayout" id="loginCmd">
+                <i class="menuIcon fa fa-right-to-bracket mx-2"></i> Se Connecter
+            </div>
+        `));
+    }
+    else
+    {
+        DDMenu.append($(`
+            <div class="dropdown-item menuItemLayout" id="logoutCmd">
+                <i class="menuIcon fa fa-right-to-bracket mx-2"></i> Se Déconnecter
+            </div>
+        `));
+    }
+    
     DDMenu.append($(`
             <div class="dropdown-divider"></div>
         `));
@@ -342,6 +377,17 @@ function updateDropDownMenu() {
     $('#loginCmd').on("click", function () {
         showLogin();
     });
+    $('#userEditCmd').on("click", function () {
+        showEditUser();
+    });
+    $('#logoutCmd').on("click", async function () {
+        let user = JSON.parse(sessionStorage.getItem("User"));
+        console.log(user);
+        users_API.Logout(user.Id);
+        sessionStorage.removeItem("User");
+        sessionStorage.removeItem("token");
+        showLogin();
+    })
     $('#allCatCmd').on("click", async function () {
         selectedCategory = "";
         await showPosts(true);
@@ -360,37 +406,40 @@ async function attach_Posts_UI_Events_Callback() {
         // Récupérer tous les likes et les appliquer correctement à chaque post
         let likes = await users_API.getAllLike();
         let user = JSON.parse(sessionStorage.getItem("User"));
-    
+        if(user != null)
+        {
+            $(".post").each(function () {
+                let postId = $(this).attr("id");  // Assurez-vous que l'ID du post est bien attribué
+                let nblikes = likes.filter(like => like.idPost == postId).length;
+                $(`#nbLikes-${postId}`).text(nblikes);  // Mettre à jour le compteur de likes
+        
+                // Vérifier si l'utilisateur a déjà liké ce post
+                let likedByUser = likes.some(like => like.idPost == postId && like.UserName == user.Name);
+                
+                // Appliquer la bonne classe (fa-solid si l'utilisateur a liké, sinon fa-regular)
+                if (likedByUser) {
+                    $(`.likeCmd[postId=${postId}]`).removeClass("fa-regular").addClass("fa-solid");
+                } else {
+                    $(`.likeCmd[postId=${postId}]`).removeClass("fa-solid").addClass("fa-regular");
+                }
+        
+                // Afficher les noms des utilisateurs au survol du nombre de likes
+                $(`#nbLikes-${postId}`).on("mouseenter", function () {
+                    let likedUsers = likes.filter(like => like.idPost == postId).map(like => like.UserName);
+                    // Créer une chaîne de texte avec les noms des utilisateurs
+                    let userList = likedUsers.join(', ');
+        
+                    // Afficher les utilisateurs dans une info-bulle ou une boîte de texte
+                    $(this).attr("title", `Liked by: ${userList}`);  // Utiliser l'attribut title pour afficher le tooltip
+                });
+        
+                $(`#nbLikes-${postId}`).on("mouseleave", function () {
+                    $(this).removeAttr("title");  // Retirer le tooltip au départ de la souris
+                });
+            });
+        }
         // Mettre à jour le nombre de likes et la classe pour chaque post
-        $(".post").each(function () {
-            let postId = $(this).attr("id");  // Assurez-vous que l'ID du post est bien attribué
-            let nblikes = likes.filter(like => like.idPost == postId).length;
-            $(`#nbLikes-${postId}`).text(nblikes);  // Mettre à jour le compteur de likes
-    
-            // Vérifier si l'utilisateur a déjà liké ce post
-            let likedByUser = likes.some(like => like.idPost == postId && like.UserName == user.Name);
-            
-            // Appliquer la bonne classe (fa-solid si l'utilisateur a liké, sinon fa-regular)
-            if (likedByUser) {
-                $(`.likeCmd[postId=${postId}]`).removeClass("fa-regular").addClass("fa-solid");
-            } else {
-                $(`.likeCmd[postId=${postId}]`).removeClass("fa-solid").addClass("fa-regular");
-            }
-    
-            // Afficher les noms des utilisateurs au survol du nombre de likes
-            $(`#nbLikes-${postId}`).on("mouseenter", function () {
-                let likedUsers = likes.filter(like => like.idPost == postId).map(like => like.UserName);
-                // Créer une chaîne de texte avec les noms des utilisateurs
-                let userList = likedUsers.join(', ');
-    
-                // Afficher les utilisateurs dans une info-bulle ou une boîte de texte
-                $(this).attr("title", `Liked by: ${userList}`);  // Utiliser l'attribut title pour afficher le tooltip
-            });
-    
-            $(`#nbLikes-${postId}`).on("mouseleave", function () {
-                $(this).removeAttr("title");  // Retirer le tooltip au départ de la souris
-            });
-        });
+        
 
     // attach icon command click event callback
     $(".editCmd").off();
@@ -662,6 +711,7 @@ function renderLoginForm() {
     $("#form").empty();
     $("#form").append($(`
         <form class="loginForm">
+            <p>${sessionStorage.getItem("messageCode")}</p>
             <div class="form-section">
                 <input type="text" id="Email" placeholder="Courriel" name="Email" required RequireMessage="Veuillez entrer un courriel" class="textInput"/>
                 <input type="text" id="Password" placeholder="Mot de Passe" name="Password" required RequireMessage="Veuillez entrer le mot de passe" class="textInput"/>
@@ -675,6 +725,7 @@ function renderLoginForm() {
         </form>
     `));
     initFormValidation(); // important do to after all html injection!
+    sessionStorage.setItem("messageCode", " ");
     $('#createAccount').on("click", function () {
         renderCreateAccountForm();
     });
@@ -687,18 +738,25 @@ function renderLoginForm() {
             let user = await users_API.Login(loginInfo);
     
             if (user != null) {
-                if (user.User.VerifyCode !== "verified") {
-                    //sessionStorage.setItem("tokenTempo", user.Access_token);
-                    sessionStorage.setItem("UserIdTempo", user.User.Id);
-                    sessionStorage.setItem("Email", loginInfo.Email);
-                    sessionStorage.setItem("Password", loginInfo.Password);
-                    await users_API.Logout(user.User.Id);
-                    showConfirmLogin();
-                } else {
-                    sessionStorage.setItem("token", user.Access_token);
-                    sessionStorage.setItem("User", JSON.stringify(user.User));
-                    showPosts();
+                if(!user.User.isBlocked)
+                {
+                    if (user.User.VerifyCode !== "verified") {
+                        //sessionStorage.setItem("tokenTempo", user.Access_token);
+                        sessionStorage.setItem("UserIdTempo", user.User.Id);
+                        sessionStorage.setItem("Email", loginInfo.Email);
+                        sessionStorage.setItem("Password", loginInfo.Password);
+                        await users_API.Logout(user.User.Id);
+                        showConfirmLogin();
+                    } else {
+                        sessionStorage.setItem("token", user.Access_token);
+                        sessionStorage.setItem("User", JSON.stringify(user.User));
+                        showPosts();
+                    }
                 }
+                else {
+                    $("#errorMessage").text("Votre compte à été bloquer!").show();
+                }
+                
             } else {
                 //console.log(users_API.currentHttpError);
                 // Si `users_API.Login` retourne null, afficher un message d'erreur générique
@@ -824,11 +882,11 @@ function renderCreateAccountForm() {
         console.log(password);
         console.log(passwordVerification);
         if (password !== passwordVerification) {
-            alert("Les mots de passe ne correspondent pas.");
+            $("#errorMessage").text("Les mots de passe doivent être identique").show();
             return;
         }
         if (email !== emailVerification) {
-            alert("Les courriels ne correspondent pas.");
+            $("#errorMessage").text("Les Email doivent être identique").show();
             return;
         }
 
@@ -854,6 +912,121 @@ function renderCreateAccountForm() {
         let response = await users_API.Register(userObject); // Enregistrement de l'utilisateur
         if (response != null) {
             //alert("Compte créé avec succès !");
+            sessionStorage.setItem("messageCode", "Votre compte à été créé. Veuillez prendre vos courriels pour réccupérer votre code de vérifivation qui vous sera demandé lors de votre prochaine connexion.");
+            renderLoginForm(); // Redirection vers le formulaire de connexion
+        }
+        else {
+            $("#errorMessage").text(users_API.currentHttpError).show();
+        }
+    });
+}
+function renderConfirmDelete(id){
+    $("#viewTitle").text("Confirmation de compte");
+        $("#form").show();
+        $("#form").empty();
+        $("#form").append($(`
+        <form class="ConfirmationForm">
+            <div class="form-submit-section">
+                <input type="button" value="Supprimer" id="delUser" class="btn btn-primary"/>
+                <input type="button" value="Annuler" id="cancel" class="btn btn-secondary"/>
+            </div>
+        </form>
+        `));
+        $('#delUser').on("click", async function () {
+            let user = JSON.parse(sessionStorage.getItem("User"));
+            await users_API.Delete(user.Id);
+            sessionStorage.removeItem("User");
+            sessionStorage.removeItem("token");
+        });
+        $('#cancel').on("click", function () {
+            showEditUser();
+        });
+}
+function renderEditAccountForm() {
+    $("#viewTitle").text("Créer un Compte");
+    $("#form").show().empty();
+    let user = JSON.parse(sessionStorage.getItem("User"));
+    // Ajout du formulaire HTML
+    $("#form").append($(`
+        <form class="createAccountForm">
+            <div class="form-section">
+                <span>Adresse Courriel</span>
+                <input type="text" value="${user.Email}" class="Email" id="Email" placeholder="Courriel" name="Email" required RequireMessage="Veuillez entrer un courriel" class="textInput"/>
+                <input type="text" value="${user.Email}" class="Email MatchedInput" id="EmailVerification" placeholder="Vérification" name="EmailVerification" required matchedInputId="Email" RequireMessage="Veuillez confirmer le courriel" class="textInput"/>
+            </div>
+            <div class="form-section">
+                <span>Mot de Passe (optionel)</span>
+                <input type="password" id="Password" placeholder="Mot de passe" name="Password" required RequireMessage="Veuillez entrer un mot de passe" class="textInput"/>
+                <input type="password" class="MatchedInput" id="PasswordVerification" placeholder="Vérification" name="PasswordVerification" required matchedInputId="Password" RequireMessage="Veuillez confirmer le mot de passe" class="textInput"/>
+            </div>
+            <div class="form-section">
+                <span>Nom</span>
+                <input type="text" value="${user.Name}" id="Username" placeholder="Nom" name="Username" required RequireMessage="Veuillez entrer un nom" class="textInput"/>
+            </div>
+            <div class="form-section">
+                <span>Avatar</span>
+                <div class='imageUploaderContainer'>
+                    <div class='imageUploader form-section-image' 
+                            newImage='./no-avatar.png' 
+                            controlId='Image' 
+                            imageSrc='${user.Avatar}' 
+                            waitingImage="Loading_icon.gif">
+                    </div>
+                </div>
+            </div>
+            <div class="form-submit-section">
+                <input type="submit" value="Enregistrer" id="EditUser" class="btn btn-primary"/>
+                <input type="button" value="Supprimer?" id="del" class="btn btn-secondary"/>
+            </div>
+            <p id="errorMessage" style="color: red; display: none;"></p>
+        </form>
+    `));
+    initFormValidation(); // Validation regex et comportement
+    //addConflictValidation('/api/check-email', 'Email', 'commitUser'); // Validation d'unicité des emails
+    initImageUploaders();
+    $('#del').on("click", function () {
+        renderConfirmDelete(user.Id);
+    });
+    $('#EditUser').on("click", async function (e) {
+        e.preventDefault(); // Empêche le rechargement de la page
+
+        // Validation des champs (mot de passe et e-mail)
+        let password = $("#Password").val();
+        let passwordVerification = $("#PasswordVerification").val();
+        let email = $("#Email").val();
+        let emailVerification = $("#EmailVerification").val();
+        console.log(password);
+        console.log(passwordVerification);
+        if (password !== passwordVerification) {
+            $("#errorMessage").text("Les mots de passe doivent être identique").show();
+            return;
+        }
+        if (email !== emailVerification) {
+            $("#errorMessage").text("Les Email doivent être identique").show();
+            return;
+        }
+
+        // Récupération des données du formulaire
+        let name = $("#Username").val();
+        let post = getFormData($(".createAccountForm"));
+        let user = JSON.parse(sessionStorage.getItem("User"));
+        let userObject = {
+            Id: user.Id,
+            Name: name,
+            Password: password,
+            Email: email,
+            Avatar: post.Image
+        };
+        console.log(userObject);
+        let response = await users_API.Modify(userObject); // Enregistrement de l'utilisateur
+        if (response != null) {
+            await users_API.Logout(user.Id);
+            sessionStorage.removeItem("token");
+            sessionStorage.removeItem("User");
+            if(user.Email != email)
+            {
+                sessionStorage.setItem("messageCode", "Votre compte à été créé. Veuillez prendre vos courriels pour récupérer votre code de vérification qui vous sera demandé lors de votre prochaine connexion.");
+            }
             renderLoginForm(); // Redirection vers le formulaire de connexion
         }
         else {
